+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Set default image size.
 * Add custom post type to ACF post type select.
+* Add custom login Image
 
 ## 0.5.4 (2023-06-26)